+++ contactlist/urls.py
@@ -1,5 +1,5 @@
 # ContactList - CTCL 2023
-# Date: June 9, 2023 - June 21, 2023
+# Date: June 9, 2023 - June 27, 2023
 # Purpose: Main application URLs
 
 from django.urls import path
@@ -11,5 +11,6 @@ urlpatterns = [
     path("view/<str:inid>/", views.view),
     path("edit/<str:inid>/", views.edit),
     path("delete/<str:inid>/", views.delete),
-    path("settings/", views.settings)
+    path("settings/", views.settings),
+    path("settings/exportcsv/", views.exportcsv)
 ]